Initial git checking of gentech incl. river and salesman solver, original work from March 1994
This work was produced in March 1994 by the authors
as an excessive work for a homework in OOP at the FH-Bielefeld
under the supervision of Prof. Dr. Bunse.

The background research and work on the actual genetic algorithm (gentech.lib)
was mostly performed by Sven Gothel using [1] - [5]
and other articles read about the matter earlier.

Christian Mueller contributed reviewing [6],
validating the algorithm and code.

We finalized the code and bug fixes together.

Bibliographie
  [1] Holland, John H. : Genetische Algorithmen,
      Spektrum der Wissenschaft, September 1992, S.44-51
  [2] Chambon, Pierre  : Gestueckelte Gene - ein Informations-Mosaik,
      Spektrum der Wissenschaft, Juli 1981, S.104-117
  [3] Dewdney, A.K.    : Genetische Algorithmen,
      Computer-Kurzweil, Spektrum-der-Wissenschaft-Verlagsgesellschaft
      1988, ISBN 3-922508-50-2, S.68-72
  [4] Weckwerth, Guido : Zeugende Zahlen,
      MC, Oktober 1993, S.54-58, Januar 1994, S.72-75
  [5] Schader, Kuhlins : Programmieren in C++,
      Springer-Verlag Heidelberg 1993, ISBN 3-540-56524-8
  [6] Schoeneburg, Eberhard : Genetische Algorithmen und Evolutionsstrategien
      Addison-Wesley 1994, ISBN 3-89319-495-2